Deduplcate impl
